requirements.txt
pymongo
bson
python-dotenv